+++ src/hooks/index.ts
@@ -1,5 +1,5 @@
 import { useIntersectionObserver } from "./useIntersectionObserver";
 import { useIsomorphicLayoutEffect } from "./useIsomorphicLayoutEffect";
-import { useScrollLock } from "./useScrollRock";
+import { useScrollLock } from "./useScrollLock";
 
 export { useIntersectionObserver, useIsomorphicLayoutEffect, useScrollLock };
